Refactor code to separate classes into their own files
This changeset separates classes `InitOption`, `Module`, their subclasses, and the function `create_module` into their own files. The classes and functions were originally present in `starterfile.py`. This change should make it easier to maintain and evolve these classes in isolation from one another.
