+++ startout/starterfile.py
@@ -1,265 +1,15 @@
 import itertools
 import os
-import platform
-import re
-import shlex
-import shutil
-import subprocess
 import sys
-import tempfile
 from typing import TextIO
 
-import requests
 import yaml
 from dotenv import load_dotenv
 from schema import Schema, And, Or, Optional, Use
 
-
-def type_tool(type_str: str) -> type or None:
-    """
-    Return the corresponding Python type based on the input string.
-
-    :param type_str: A string representing the desired Python type.
-                     Possible values are "int", "float", "str", or "string".
-    :return: The corresponding Python type if it exists, otherwise None.
-    """
-    types = {
-        "int": int,
-        "float": float,
-        "str": str,
-        "string": str,
-    }
-    try:
-        return types[type_str.lower()]
-    except KeyError:
-        return None
-
-
-def replace_env(string: str) -> str:
-    """
-    :param string: The string in which to replace environment variable placeholders.
-    :return: The string with all environment variable placeholders replaced with their corresponding values.
-
-    This method takes a string as input and replaces all occurrences of environment variable placeholders in the
-    format ${variable_name} with their corresponding values. It uses regular expressions to find all placeholders
-    in the string, then checks if the corresponding environment variable is set. If the variable is set, it replaces
-    the placeholder with the variable's value. If the variable is not set, it raises a ValueError.
-
-    Example usage:
-
-    >>> replace_env("Hello ${USERNAME}, your home directory is ${HOME}")
-    'Hello John, your home directory is /home/john'
-    """
-    pattern = re.compile(r'\$\{(.+?)}')
-    matches = pattern.findall(string)
-
-    for match in matches:
-        env_value = os.getenv(match)
-        if env_value is None:
-            raise ValueError(f'Environment variable {match} not set.')
-        string = string.replace(f'${{{match}}}', env_value)
-
-    return string
-
-
-def run_script_with_env_substitution(script_str: str, verbose: bool = False) -> tuple[str, int]:
-    """
-    Run a script with environment variable substitution. If the script fails to run as a shlex'd list, run it as a
-    string instead.
-
-    :param verbose: Whether to print warning if the script fails to run as a shlex'd list or not.
-    :param script_str: The script to be executed as a string.
-    :return: A tuple containing the stdout output and the return code of the script.
-    """
-
-    # Inject environment variables
-    substituted_script = replace_env(script_str)
-
-    # Check if command can be found with shutil, run as a string otherwise
-    _script = shlex.split(substituted_script)
-
-    try:
-        if shutil.which(_script[0]) is None:
-            if verbose:
-                print(f"'{_script[0]}' is not installed. Trying script in shell.", file=sys.stderr)
-            result = subprocess.run(substituted_script, shell=True, check=True, text=True, capture_output=True)
-        else:
-            result = subprocess.run(_script, shell=True, check=True, text=True, capture_output=True)
-    except OSError as e:
-        return f"{e}", 1
-    except subprocess.CalledProcessError as e:
-        return f"{e.stdout}\n{e.stderr}", e.returncode
-
-    return result.stdout, result.returncode
-
-
-class Tool:
-    """
-    Class representing a tool with installation and uninstallation scripts.
-
-    Attributes:
-        tool_scripts_schema (Schema): Schema definition for tool scripts.
-        tool_schema (Schema): Schema definition for the entire tool object.
-
-    Methods:
-        __init__(self, name: str, dependencies: list[str], scripts: dict[str, str or dict[str, str]]):
-            Initialize a Tool instance with the provided name, dependencies, and scripts.
-            Raises a TypeError if 'install' or 'uninstall' scripts are not defined.
-
-        get_script(self, script: str, scripts_list, name: str or None = None) -> str or None:
-            Get the script for the given name and platform. May be run basically static, used for instantiation.
-            Returns the script string if found, else None.
-
-        run(self, script: str) -> tuple[str, int]:
-            Execute the specified script and return the output and return code.
-            Returns a tuple of the stdout and return code.
-
-        check(self) -> bool:
-            Run the 'check' script and return True if the return code is 0, else False.
-
-        initialize(self) -> bool:
-            Run the 'install' script and print the output.
-            Returns True if the return code is 0, else False.
-
-        destroy(self) -> bool:
-            Run the 'uninstall' script and print the output.
-            Returns True if the return code is 0, else False.
-    """
-    tool_scripts_schema = Schema(
-        Or(
-            {
-                Optional("install"): And(str),
-                Optional("uninstall"): And(str),
-                Optional("check"): And(str),
-                Optional("windows"): {
-                    Optional("install"): And(str),
-                    Optional("uninstall"): And(str),
-                    Optional("check"): And(str),
-                },
-                Optional("mac"): {
-                    Optional("install"): And(str),
-                    Optional("uninstall"): And(str),
-                    Optional("check"): And(str),
-                },
-                Optional("linux"): {
-                    Optional("install"): And(str),
-                    Optional("uninstall"): And(str),
-                    Optional("check"): And(str),
-                },
-            },
-        )
-    )
-    tool_schema = Schema(
-        {
-            Optional("depends_on"): Or(str, list[str]),
-            "scripts": tool_scripts_schema
-        }
-    )
-
-    def __init__(self, name: str, dependencies: list[str], scripts: dict[str, str or dict[str, str]]):
-        """
-        Initializes a Tool with the given name, dependencies, and scripts.
-
-        :param name: The name of the tool.
-        :param dependencies: A list of dependencies required by the tool.
-        :param scripts: A dictionary mapping script names to their respective commands or scripts.
-
-        :raises TypeError: If the 'install' and 'uninstall' scripts are not defined for the module.
-        """
-        if self.get_script("install", scripts, name=name) is None:
-            raise TypeError(f"No 'install' script defined for module \"{name}\". Failed to create Module.")
-        if self.get_script("uninstall", scripts, name=name) is None:
-            raise TypeError(f"No 'uninstall' script defined for module \"{name}\". Failed to create Module.")
-
-        self.name = name
-        self.dependencies = dependencies
-        self.scripts = scripts
-
-    def get_script(self, script: str, scripts_dict: dict[str, str] or None = None,
-                   name: str or None = None) -> str or None:
-        """
-        Retrieve the script based on the provided script name.
-
-        :param script: The name of the script to retrieve.
-        :param scripts_dict: A dictionary containing the available scripts. If not provided, use the default scripts.
-        :param name: The name of the tool. If not provided, use the default tool name.
-        :return: The script content as a string, or None if the script does not exist.
-        :raises ValueError: If the provided script name does not exist in the scripts_dict.
-        """
-        if name is None:
-            name = self.name
-        if scripts_dict is None:
-            scripts_dict = self.scripts
-
-        _os = platform.system().lower()
-        windows = _os in ["windows", "win32"]
-        macos = _os in ["darwin"]
-
-        _script = None
-
-        # Default to top-level definition of the script (not platform-dependent)
-        if script in scripts_dict:
-            _script = scripts_dict[script]
-
-        # Any platform-dependent scripts will override the top-level definition
-        if windows and "windows" in scripts_dict.keys():
-            if script in scripts_dict["windows"]:
-                _script = scripts_dict["windows"][script]
-        elif macos and "mac" in scripts_dict.keys():
-            if script in scripts_dict["mac"]:
-                _script = scripts_dict["mac"][script]
-        elif (not windows and not macos) and "linux" in scripts_dict.keys():
-            if script in scripts_dict["linux"]:
-                _script = scripts_dict["linux"][script]
-        else:
-            if _script is None:
-                raise ValueError(f"Tool \"{name}\" does not have script '{script}' "
-                                 f"in {list(scripts_dict.keys())}")
-
-        return _script
-
-    def run(self, script: str) -> tuple[str, int]:
-        """
-        Runs a script with environment variable substitutions.
-
-        :param script: The name of the script to be executed, located in the Tool's scripts.
-        :return: A tuple containing the stdout output and the return code of the script execution.
-        """
-        _script = self.get_script(script)
-
-        return run_script_with_env_substitution(_script)
-
-    def check(self):
-        """
-        Run the Tool's 'check' script.
-
-        :return: True if the response code is 0, False otherwise.
-        """
-        response, code = self.run("check")
-
-        return code == 0
-
-    def initialize(self):
-        """
-        Run the Tool's 'install' script.
-
-        :return: True if the response code is 0, False otherwise.
-        """
-        msg, code = self.run("install")
-        print(msg)
-
-        return code == 0
-
-    def destroy(self):
-        """
-        Run the Tool's 'uninstall' script.
-
-        :return: True if the response code is 0, False otherwise.
-        """
-        msg, code = self.run("uninstall")
-        print(msg)
-
-        return code == 0
+from startout.module import Module, create_module
+from startout.tool import Tool
+from startout.util import replace_env, type_tool
 
 
 class InitOption:
@@ -297,244 +47,6 @@ class InitOption:
         self.default = replace_env(default) if type(default) is str else default
         self.prompt = replace_env(options_set["prompt"])
         self.value = None
-
-
-class Module:
-    """
-    Class representing a module.
-
-    Module.source defines how the module is collected, defined as:
-
-    - git
-      : A URI is passed to `git clone`
-
-    - script
-      : A script is executed by `bash` (or `git-bash` on Windows)
-
-    Attributes:
-        name (str): The name of the module.
-        dest (str): The destination of the module (usually used as the name of the directory into which the module is installed).
-        source (dict): The source of the module, given as any ONE of [git, curl, script, docker].
-        scripts (dict): Scripts associated with the module.
-        dependencies (str or list[str]): Dependencies of the module. (Optional)
-        init_options (list[dict]): Initialization options for the module. (Optional)
-
-    """
-    module_schema = Schema(
-        {
-            "dest": And(str, Use(replace_env)),
-            "source": Schema(
-                {
-                    Or("git", "script", only_one=True): str
-                }
-            ),
-            "scripts": And(dict, len),
-            Optional("depends_on"): Or(str, list[str]),
-            Optional("init_options"): list[Schema(
-                {
-                    "env_name": And(str, len),
-                    "type": And(str, len),
-                    "default": And(str, len),
-                    "prompt": And(str, len),
-                }
-            )]
-        }
-    )
-
-    def __init__(self, name: str, dest: str, source: str, scripts: dict[str, str], dependencies=None,
-                 init_options=None):
-        """
-        Initialize a new Module instance.
-
-        :param name: The name of the module.
-        :param dest: The destination path of the module.
-        :param source: The source path of the module.
-        :param scripts: A dictionary mapping script names to script paths.
-        :param dependencies: (optional) A list of module names that this module depends on. Defaults to None.
-        :param init_options: (optional) Additional options for module initialization. Defaults to None.
-        """
-        if "init" not in scripts.keys():
-            raise TypeError(f"No 'init' script defined for module \"{name}\". Failed to create Module.")
-        if "destroy" not in scripts.keys():
-            raise TypeError(f"No 'destroy' script defined for module \"{name}\". Failed to create Module.")
-
-        self.name = name
-        self.dest = dest
-        self.source = source
-        self.scripts = scripts
-        self.dependencies = dependencies
-        self.init_options = init_options
-
-    def run(self, script: str, print_output: bool = False) -> tuple[str, int]:
-        """
-        Runs a script with environment variable substitutions.
-
-        :param script: The name of the script to be executed, located in the Module's scripts.
-        :param print_output: Whether to print the response at the .... level
-        :return: A tuple containing the stdout output and the return code of the script execution.
-        """
-        if script not in self.scripts:
-            raise ValueError(f"Module \"{self.name}\" does not have script '{script}' in {list(self.scripts.keys())}")
-
-        response, code = run_script_with_env_substitution(self.scripts[script])
-
-        if print_output and len(response.strip()) > 0:
-            print(f".... [{self.name}.{script}]: {response.strip()}")
-
-        return response, code
-
-    def initialize(self):
-        """
-        Run the Tool's 'init' script.
-
-        :return: True if the response code is 0, False otherwise.
-        """
-        msg, code = self.run("init", print_output=True)
-
-        if code != 0:
-            print(f".. FAILURE [{self.name}]: {msg}", file=sys.stderr)
-            return False
-        else:
-            print(f".. SUCCESS [{self.name}]: Initialized module {self.name}")
-            return True
-
-    def destroy(self):
-        """
-        Run the Tool's 'destroy' script.
-
-        :return: True if the response code is 0, False otherwise.
-        """
-        msg, code = self.run("destroy", print_output=True)
-
-        if code != 0:
-            print(f".. FAILURE [{self.name}]: {msg}", file=sys.stderr)
-            return False
-        else:
-            print(f".. SUCCESS [{self.name}]: Destroyed module {self.name}")
-            return True
-
-
-class GitModule(Module):
-    """
-    Module for interacting with Git.
-
-    :class:`GitModule` is a subclass of :class:`Module` and provides functionality
-    for initializing a Git repository.
-
-    Example:
-        >>> git_module = GitModule()
-        >>> git_module.source = "https://github.com/username/repo.git"
-        >>> git_module.dest = "/path/to/destination"
-        >>> git_module.initialize()
-
-    Note:
-        This module requires Git to be installed on the system.
-
-    Attributes:
-        source (str): The source URL of the Git repository.
-        dest (str): The destination directory to clone the repository into.
-
-    Raises:
-        OSError: If Git is not installed on the system.
-
-    """
-    def initialize(self):
-        """
-        Initializes the object by cloning `self.source` (a repository) to `self.dest` (a directory).
-
-        :return: bool - True if the cloning succeeds, False otherwise.
-        :raises OSError: If Git is not installed.
-        """
-        if shutil.which("git") is None:
-            raise OSError(f"Git is not installed. Please install Git and try again.")
-
-        cmd = [
-            "git",
-            "clone",
-            "--progress",
-            self.source,
-            self.dest
-        ]
-        result = subprocess.run(cmd, shell=True, check=True, text=True, capture_output=True)  # TODO hoist the feedback to the terminal, especially if it can be displayed by the CLI which enwraps it
-
-        if result.returncode != 0:
-            print(f".. FAILURE [{self.name}]: {result.stdout.strip()}", file=sys.stderr)
-            return False
-        else:
-            print(f".... PROGRESS [{self.name}]: Cloned module {self.name}, running init script")
-            msg, code = self.run("init", print_output=True)
-
-            if code != 0:
-                print(f".. FAILURE [{self.name}]: {msg}", file=sys.stderr)
-                return False
-            else:
-                print(f".. SUCCESS [{self.name}]: Initialized module {self.name}")
-                return True
-
-
-class ScriptModule(Module):
-    def initialize(self):
-        _os = platform.system().lower()
-        windows = _os in ["windows", "win32"]
-
-        if windows and shutil.which("git-bash") is None:
-            raise OSError(f"Windows detected and Git Bash is not installed. Please install Git Bash and try again.")
-
-        _script = f"{'git-bash' if windows else 'bash'} -c \"{self.source}\""
-
-        response, code = run_script_with_env_substitution(_script, verbose=True)
-
-        if code != 0:
-            print(f".. FAILURE [{self.name}]: {response}", file=sys.stderr)
-            return False
-        else:
-            msg = response.strip()
-            if len(msg) > 0:
-                print(f".... [{self.name}.source.script]: {response.strip()}")
-            print(f".. SUCCESS [{self.name}]: Ran script for module {self.name}")
-            return True
-
-
-def create_module(module: dict, name: str):
-    """
-    Create a module object based on the given parameters. The value of module.source determines the type of module.
-
-    :param module: A dictionary representing the module information.
-    :param name: The name of the module.
-    :return: An instance of a module object.
-    """
-    mode = next(iter(module["source"]))
-    source = module["source"][mode]
-    dest = module["dest"]
-
-    options = None
-    if "init_options" in module.keys():
-        options = []
-        for options_set in module["init_options"]:
-            options.append(InitOption(options_set))
-    dependencies = None
-    if "depends_on" in module.keys():
-        _deps = module["depends_on"]
-        if type(_deps) is str:
-            dependencies = [_deps]
-        else:
-            dependencies = _deps
-
-    # Instantiate the correct type of Module
-    _T = Module
-    if mode == "git":
-        _T = GitModule
-    elif mode == "script":
-        _T = ScriptModule
-
-    return _T(
-        name=name,
-        dest=dest,
-        source=source,
-        scripts=module["scripts"],
-        init_options=options,
-        dependencies=dependencies
-    )
 
 
 class Starter:
